backend/requirements.txt
# Backend Dependencies - Lake Holidays Challenge
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-oauth2==1.1.1
httpx-oauth==0.16.1

# AI/LLM Integration
openai==1.3.8
azure-ai-textanalytics==5.3.0
azure-storage-blob==12.19.0

# Image Processing
Pillow==10.1.0
python-magic==0.4.27

# Geolocation & Maps
geopy==2.4.1
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# CORS and middleware
#fastapi-cors==0.0.6

# Background tasks
celery==5.3.4
redis==5.0.1
